requirements.txt
scipy
mlflow
scikit-learn
seaborn
matplotlib
folium
geopandas